Refactor RetailProduct and related classes to use Long IDs instead of int, and update User and UserCredential classes for consistency
- Updated RetailProductFacadeImpl to use user ID instead of user sequence.
- Refactored RetailProductMapper to align with new RetailProduct structure.
- Modified AuthToken to replace user sequence with user ID.
- Changed EntityType, PaymentMethod, RetailCategory, TaxGroup, and TaxRate to use long for IDs.
- Refactored RetailProduct to include new fields and removed unnecessary ones.
- Introduced RetailProductBarcode, RetailProductCost, RetailProductMaster, RetailProductMasterPolicy, and RetailProductPrice classes.
- Removed UserReference class and updated User and UserCredential classes to reflect changes in ID handling.
- Updated GenericEntity to support Long IDs and adjusted constructors accordingly.
- Fixed RetailProductNumberGenerator to return the correct product number.
- Updated ConstantEntity interface to use long for ID.

+++ src/main/java/com/lesconstructionssapete/stempyerp/core/domain/base/constant/EntityType.java
@@ -4,14 +4,14 @@ import com.lesconstructionssapete.stempyerp.core.shared.constant.ConstantEntity;
 
 public class EntityType implements ConstantEntity {
 
-  private final int id;
+  private final long id;
   private final String name;
   private final int padLength;
   private final String prefixString;
   private final boolean enabled;
 
   public EntityType(
-      int id,
+      long id,
       String name,
       int padLength,
       String prefixString,
@@ -24,7 +24,7 @@ public class EntityType implements ConstantEntity {
   }
 
   @Override
-  public int getId() {
+  public long getId() {
     return id;
   }
